+++ ovp_search/filters.py
@@ -1,10 +1,65 @@
-from rest_framework.filters import OrderingFilter
 from ovp_search import helpers
 from haystack.query import SQ
+
+from rest_framework.filters import OrderingFilter
+from rest_framework.exceptions import NotAuthenticated
+
+from django.db.models import When, F, IntegerField, Count, Case
+
 import json
 
-class SearchOrderingFilter(OrderingFilter):
-  pass
+#####################
+## ViewSet filters ##
+#####################
+
+class ProjectRelevanceOrderingFilter(OrderingFilter):
+  def get_skills_and_causes(self, request):
+    if not request.user.is_authenticated():
+      raise NotAuthenticated()
+
+    user = request.user
+    output = {"skills": [], "causes": []}
+
+    if user.profile:
+      output["skills"] = user.profile.skills.values_list('id', flat=True)
+      output["causes"] = user.profile.causes.values_list('id', flat=True)
+
+    return output
+
+  def annotate_queryset(self, queryset, request):
+    self.get_skills_and_causes(request)
+
+    queryset = queryset\
+                .annotate(\
+                  cause_relevance =
+                    Count(
+                      Case(When(causes__pk__in=[1,3], then=1),
+                           output_field=IntegerField()),
+                      distinct=True),
+                  skill_relevance =
+                    Count(
+                      Case(When(skills__pk__in=[1,2,4], then=1),
+                                output_field=IntegerField()),
+                      distinct=True))\
+                .annotate(relevance = F('cause_relevance') + F('skill_relevance'))
+
+    return queryset
+
+  def filter_queryset(self, request, queryset, view):
+    ordering = self.get_ordering(request, queryset, view)
+
+    if "relevance" in ordering or "-relevance" in ordering:
+      queryset = self.annotate_queryset(queryset, request)
+
+    if ordering:
+      return queryset.order_by(*ordering)
+
+    return queryset
+
+
+######################
+## Haystack filters ##
+######################
 
 def by_skills(queryset, skill_string=''):
   """ Filter queryset by a comma delimeted skill list """