+++ CHANGES.rst
@@ -94,3 +94,4 @@ v1.0.13
 v1.0.14[unreleased]
 -----------
 * Remove 'ordered' param in favor of OrderingFilter
+* Implement project ordering by 'relevance'